customer_churn_analysis/requirements.txt
sklearn
pandas
seaborn
matplotlib
dvc
google-api-python-client
dask[dataframe]
gcsfs
google-cloud
google-cloud-storage
